fix: move all modules to live_divulgator
