+++ live_divulgator/helpers/slice.py
