+++ live_divulgator/plugins/plugin.py
