+++ live_divulgator/twitch/exceptions.py
