+++ test/bot/test_bot.py
@@ -1,5 +1,5 @@
 from pytest import fixture
-from src.bot.live_divulgator import LiveDivulgator
+from src.bot.bots.live_divulgator import LiveDivulgator
 from src.bot.plugins.twitter import TwitterPlugin
